refactor: remove repo link and related console log
The repo link and its associated console log in the root component have been removed as they are no longer relevant. The link was previously used to point to the GitHub repository, but it has been deprecated. The console log was used to display the repository link, which is no longer necessary.